refactor

+++ MailingService/Entities/RabbitMqSettings.cs
@@ -6,5 +6,5 @@ public class RabbitMqSettings
     public string Username { get; init; } = string.Empty;
     public string Password { get; init; } = string.Empty;
     public string QueueName { get; init; } = string.Empty;
-    public int KillSwitchTimeoutMinutes { get; init; }
+    public float KillSwitchTimeoutMinutes { get; init; }
 }
